[DT][NFC] Localize CPU specific encoding materialization logic.
The revision moves the CPU materialization logic from
Dialect/Codegen/Utils/Utils.[h|cpp] to CPUEncodingExternalModels. They
were public methods during transition states. After all the CPU
layout attributes are implemented, we no longer need to expose them to
public.

Additionally, it removes the outdated logic from MaterializeContraction
pattern. And it removes the `transposeNarrowN` input argument from
lowerContractionOpWithEncoding method because all the CPU backends
enable the transposeNarrowN feature.
