+++ compiler/src/iree/compiler/Codegen/Dialect/Codegen/Utils/Utils.h
@@ -75,29 +75,6 @@ struct TileMxNxK {
 MaterializeEncodingInfo
 getEncodingInfoForMatmul(Encoding::EncodingAttr encoding, TileMxNxK tileMxNxK);
 
-//===----------------------------------------------------------------------===//
-// Operation Lowering Utilities.
-//===----------------------------------------------------------------------===//
-
-// TODO(hanchung): The below methods are exposed to public because they are
-// shared between MaterializeEncodingIntoPackUnPack.cpp.cpp and
-// CPUEncodingExternalModels.cpp. They will be moved to other places after all
-// the CPU backends implement their layout attributes.
-
-/// Returns the best TileMxNxK from `enumeratedTiles` pool. If the
-/// `hostDefinedUpperBound` is not empty, the chosen tile sizes can not be
-/// greater than the values.
-/// TODO(#16933): Remove `hostDefinedUpperBound` once we can propagate such
-/// information to host. For now, they are defined by host.
-TileMxNxK chooseMatmulTile(ArrayRef<TileMxNxK> enumeratedTiles,
-                           IREE::Encoding::MatmulNarrowDim narrowDim,
-                           ArrayRef<int64_t> hostDefinedUpperBound = {});
-
-FailureOr<Operation *>
-lowerContractionOpWithEncoding(OpBuilder &builder, linalg::LinalgOp linalgOp,
-                               ValueRange operands, bool transposeNarrowN,
-                               LayoutAttrInterface layoutAttr);
-
 } // namespace mlir::iree_compiler::IREE::Codegen
 
 #endif // IREE_COMPILER_CODEGEN_DIALECT_CODEGEN_UTILS_H_
